backend-assignment/requirements.txt
python-dotenv
google-generativeai
google-cloud-texttospeech
google-cloud-speech
pyaudio
pydub
